requirements.txt
setuptools
numpy
bob.extension
bob.db.drive
bob.db.stare
bob.db.chasedb1
bob.db.hrf
bob.db.drionsdb
bob.db.rimoner3
bob.db.drishtigs1
bob.db.refuge
bob.db.iostar
pytorch >=1.0.1
torchvision
pandas
matplotlib